requirements.txt
crewai
python-dotenv
requests 